Switch 1(Zrealizowac platnosc) zrobiony na full, dziala wczytywanie danych, zapisywanie wpisow do archiwum, zapisywanie nowych danych do plikow z informacjami o osobach i ich saldach

+++ src/com/github/mateuszhorczak/Karta.java
@@ -16,7 +16,7 @@ public abstract class Karta {
         return stanKarty;
     }
 
-    public boolean platnosc(int kwota){
+    public boolean platnosc(double kwota){
         if (stanKarty >= kwota) {
             stanKarty = stanKarty - kwota;
             return true;
